feat:Updated all the changes in the API Endpoint

+++ packages/Webkul/AdminApi/src/ApiDataSource/Catalog/CategoryDataSource.php
@@ -5,6 +5,7 @@ namespace Webkul\AdminApi\ApiDataSource\Catalog;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Webkul\AdminApi\ApiDataSource;
 use Webkul\Category\Repositories\CategoryRepository;
+use Webkul\Core\Repositories\ChannelRepository;
 
 class CategoryDataSource extends ApiDataSource
 {
@@ -20,7 +21,10 @@ class CategoryDataSource extends ApiDataSource
      *
      * @return void
      */
-    public function __construct(protected CategoryRepository $categoryRepository) {}
+    public function __construct(
+        protected CategoryRepository $categoryRepository,
+        protected ChannelRepository $channelRepository,
+    ) {}
 
     /**
      * Prepares the query builder for API requests.
@@ -112,18 +116,23 @@ class CategoryDataSource extends ApiDataSource
                 ],
             ],
         ];
+
         $category = $this->processRequestedFilters($requestedFilters)->first();
+
         if (! $category) {
             throw new ModelNotFoundException(
                 trans('admin::app.catalog.categories.not-found', ['code' => (string) $code])
             );
         }
+
+        if ($this->isRelatedToChannel($category->id)) {
+
+            throw new \Exception(trans('admin::app.catalog.categories.delete-category-root'));
+        }
+
         $category->delete();
 
-        return [
-            'message' => trans('admin::app.catalog.categories.delete-success'),
-            'code'    => $category['code'],
-        ];
+        return true;
     }
 
     /**
